+++ lib/flop/meta.ex
@@ -1,0 +1,50 @@
+defmodule Flop.Meta do
+  @moduledoc """
+  Defines a struct for holding meta information of a query result.
+  """
+
+  @typedoc """
+  Meta information for a query result.
+  """
+  @type t :: %__MODULE__{
+          current_offset: non_neg_integer,
+          current_page: pos_integer,
+          has_next_page?: boolean,
+          has_previous_page?: boolean,
+          next_offset: non_neg_integer | nil,
+          next_page: pos_integer | nil,
+          page_size: pos_integer | nil,
+          previous_offset: non_neg_integer | nil,
+          previous_page: pos_integer | nil,
+          total_count: non_neg_integer,
+          total_pages: non_neg_integer
+        }
+
+  @enforce_keys [
+    :current_offset,
+    :current_page,
+    :has_next_page?,
+    :has_previous_page?,
+    :next_offset,
+    :next_page,
+    :page_size,
+    :previous_offset,
+    :previous_page,
+    :total_count,
+    :total_pages
+  ]
+
+  defstruct [
+    :current_offset,
+    :current_page,
+    :has_next_page?,
+    :has_previous_page?,
+    :next_offset,
+    :next_page,
+    :page_size,
+    :previous_offset,
+    :previous_page,
+    :total_count,
+    :total_pages
+  ]
+end